+++ src/components/Buttons/RemoveBtn.tsx
@@ -1,0 +1,7 @@
+interface IRemoveBtn {
+    cb: any
+}
+
+export default function RemoveBtn({ cb }: IRemoveBtn) {
+    return <i className="btn fas fa-trash-alt" onClick={() => cb()}></i>
+}