+++ src/components/WidgetTabAdd.tsx
@@ -1,14 +1,8 @@
-import React,{ useState } from "react";
+import React, { useState } from "react";
 import store from "../store";
 import Modal from "./Modal";
-import { widgets } from "../utils/widgetOptions";
 import { observer } from "mobx-react-lite";
-import Watchlist from "./Widgets/Watchlist";
-import Matrix from "./Widgets/Matrix";
-import Screener from "./Widgets/Screener";
-import Chart from "./Widgets/Chart";
-import PositionTracker from "./Widgets/PositionTracker";
-import Wallets from "./Widgets/Wallets";
+import { IOptions, widgets } from "./Widget";
 
 interface IWidgetTabAdd {
     widgetID: number;
@@ -30,27 +24,27 @@ function WidgetTabAdd({ widgetID }: IWidgetTabAdd) {
     const WidgetOptions = () => {
         const handleClick = (widget: string) => {
             switch (widget) {
-                case "Screener":
-                    store.updateWidget(widgetID, { id: 1, type: 'screener', tab: 'Market Screener', isActive: false, content: <Screener /> });
+                case "Market Screener":
+                    store.updateWidget(widgetID, widgets.filter((e: IOptions) => e.tab === "Market Screener")[0]);
                     break;
                 case "Chart":
-                    store.updateWidget(widgetID, { id: 2, type: 'chart', tab: 'Chart', isActive: false, content: <Chart /> });
+                    store.updateWidget(widgetID, widgets.filter((e: IOptions) => e.tab === "Chart")[0]);
                     break;
                 case 'Watchlist':
-                    store.updateWidget(widgetID, { id: 3, type: 'watchlist', tab: 'Watchlist', isActive: false, content: <Watchlist /> });
+                    store.updateWidget(widgetID, widgets.filter((e: IOptions) => e.tab === "Watchlist")[0]);
                     break;
                 case 'Matrix':
-                    store.updateWidget(widgetID, { id: 4, type: 'matrix', tab: 'Matrix', isActive: false, content: <Matrix /> });
+                    store.updateWidget(widgetID, widgets.filter((e: IOptions) => e.tab === "Matrix")[0]);
                     break;
                 case 'Position Tracker':
-                    store.updateWidget(widgetID, { id: 5, type: 'positionTracker', tab: 'Position Tracker', isActive: false, content: <PositionTracker /> });
+                    store.updateWidget(widgetID, widgets.filter((e: IOptions) => e.tab === "Position Tracker")[0]);
                     break;
                 case 'Wallets':
-                    store.updateWidget(widgetID, { id: 6, type: 'wallets', tab: 'Wallets', isActive: false, content: <Wallets /> });
+                    store.updateWidget(widgetID, widgets.filter((e: IOptions) => e.tab === "Wallets")[0]);
                     break;
             }
         }
-        return <>{widgets.map((e: string, i: number) => <p key={i} onClick={() => handleClick(e)}>{e}</p>)}</>
+        return <>{widgets.map((e: IOptions, i: number) => <p key={i} onClick={() => handleClick(e.tab)} className={'btn'}>{e.tab}</p>)}</>
     }
 
     if (toggle || store.widgetBarHoverd) {