+++ src/data/index.ts
@@ -5,8 +5,8 @@ const friends: any[] = require("./friends");
 export { records, watchlist, friends };
 
 interface IAddToWatchlist {
-    symbol: string;
-    list: string;
+  symbol: string;
+  list: string;
 }
 
 export const getWatchlist = () =>
@@ -17,12 +17,16 @@ export const getWatchlist = () =>
     };
   });
 
-export const addToWatchlist = ({symbol, list}: IAddToWatchlist) => {
-  console.log({symbol, list});
-  
-  watchlist.push({symbol, list});
+export const addToWatchlist = ({ symbol, list }: IAddToWatchlist) => {
+  watchlist.push({ symbol, list });
   return getWatchlist();
 };
 
 export const getWatchlistNames = () =>
   Array.from(new Set(watchlist.map((e) => e.list)));
+
+export const removeFromWatchlist = ({ symbol, list }: IAddToWatchlist) => {
+  const index = watchlist.findIndex((e) => e.symbol === symbol && e.list === list);
+  watchlist.splice(index, 1);
+  return getWatchlist();
+}