+++ src/utils/helpers.ts
@@ -17,12 +17,28 @@ export const widgetCount = (widgets: number) => {
     case 3:
     case 4:
       return 2;
-    
+
     case 5:
     case 6:
       return 3;
-  
+
     default:
       return widgets;
   }
+}
+
+// debounce function
+export const debounce = (func: Function, wait: number, immediate: boolean = false) => {
+  let timeout: any;
+  return function (this: any) {
+    let context = this, args = arguments;
+    let later = function () {
+      timeout = null;
+      if (!immediate) func.apply(context, args);
+    };
+    let callNow = immediate && !timeout;
+    clearTimeout(timeout);
+    timeout = setTimeout(later, wait);
+    if (callNow) func.apply(context, args);
+  };
 }