+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0.0.6
+- Added feature "Drag and Drop" between watchlists
+- Added remove button for watchlist
+- Added filters to screener, filters appears when widget is maximized
+
 ## v0.0.5
 - Added feature "Drag and Drop", now is possible to drag a symbol from Market Screener and drop in Watchlist